+++ App.tsx
@@ -1,7 +1,10 @@
 import React, {useEffect} from 'react';
+import {StyleSheet} from 'react-native';
+import {GestureHandlerRootView} from 'react-native-gesture-handler';
 import {NavigationContainer} from '@react-navigation/native';
 import {SafeAreaProvider} from 'react-native-safe-area-context';
 import {ThemeProvider} from './src/app/providers/ThemeProvider';
+import {NotificationAttentionProvider} from './src/app/providers/NotificationAttentionProvider';
 import {RootNavigator} from './src/app/navigation/RootNavigator';
 import {navigationRef} from './src/app/navigation/navigationRef';
 import {
@@ -27,14 +30,24 @@ function App(): React.JSX.Element {
   useEffect(() => onUnlockPress(navigateToUnlock), []);
 
   return (
-    <SafeAreaProvider>
-      <ThemeProvider>
-        <NavigationContainer ref={navigationRef}>
-          <RootNavigator />
-        </NavigationContainer>
-      </ThemeProvider>
-    </SafeAreaProvider>
+    <GestureHandlerRootView style={styles.root}>
+      <SafeAreaProvider>
+        <ThemeProvider>
+          <NotificationAttentionProvider>
+            <NavigationContainer ref={navigationRef}>
+              <RootNavigator />
+            </NavigationContainer>
+          </NotificationAttentionProvider>
+        </ThemeProvider>
+      </SafeAreaProvider>
+    </GestureHandlerRootView>
   );
 }
+
+const styles = StyleSheet.create({
+  root: {
+    flex: 1,
+  },
+});
 
 export default App;