+++ src/app/navigation/types.ts
@@ -1,14 +1,22 @@
 import type {NativeStackScreenProps} from '@react-navigation/native-stack';
+import type {CompositeScreenProps} from '@react-navigation/native';
+import type {DrawerScreenProps} from '@react-navigation/drawer';
 
 export type RootStackParamList = {
   Home: undefined;
   CreateMessage: undefined;
   RevealMessage: {messageId: string};
+  Login: undefined;
+  Settings: undefined;
 };
 
-export type HomeScreenProps = NativeStackScreenProps<
-  RootStackParamList,
-  'Home'
+export type HomeDrawerParamList = {
+  HomeMain: undefined;
+};
+
+export type HomeScreenProps = CompositeScreenProps<
+  DrawerScreenProps<HomeDrawerParamList, 'HomeMain'>,
+  NativeStackScreenProps<RootStackParamList>
 >;
 export type CreateMessageScreenProps = NativeStackScreenProps<
   RootStackParamList,
@@ -17,4 +25,12 @@ export type CreateMessageScreenProps = NativeStackScreenProps<
 export type RevealMessageScreenProps = NativeStackScreenProps<
   RootStackParamList,
   'RevealMessage'
+>;
+export type LoginScreenProps = NativeStackScreenProps<
+  RootStackParamList,
+  'Login'
+>;
+export type SettingsScreenProps = NativeStackScreenProps<
+  RootStackParamList,
+  'Settings'
 >;